requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.21.0
Pillow>=9.0.0
scikit-learn>=1.0.0
python-multipart>=0.0.5

